Use intptr_t/size_t instead of uint64_t

+++ core/allocator.hpp
@@ -1,21 +1,22 @@
 #pragma once
 
+#include "stddef.h"
 #include "stdint.h"
 
 class Allocator {
 public:
     static Allocator& global();
 
-    void add_region(uint64_t start, uint64_t size);
+    void add_region(uintptr_t start, size_t size);
 
-    void* alloc(uint64_t size);
+    void* alloc(size_t size);
     void dealloc(void* addr);
 
     void dump() const;
 
 private:
     struct FreeChunk {
-        uint64_t size;
+        size_t size;
         FreeChunk* next;
     };
