+++ core/page_allocator.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "stddef.h"
 #include "stdint.h"
 #include "list.hpp"
 
@@ -7,21 +8,21 @@ class PageAllocator {
 public:
     static PageAllocator& global();
 
-    void add_region(uint64_t start, uint64_t size);
-    void reserve_region(uint64_t start, uint64_t size);
+    void add_region(uintptr_t start, size_t size);
+    void reserve_region(uintptr_t start, size_t size);
 
-    uint64_t alloc(uint64_t size);
+    uintptr_t alloc(size_t size);
 
     void dump() const;
 private:
     struct PageChunk {
-        uint64_t start;
-        uint64_t size;
+        uintptr_t start;
+        size_t size;
     };
 
     List<PageChunk> m_huge_chunks;
     List<PageChunk> m_big_chunks;
     List<PageChunk> m_chunks;
 
-    void add_chunk(List<PageChunk>& chunks, uint64_t start, uint64_t size);
+    void add_chunk(List<PageChunk>& chunks, uintptr_t start, size_t size);
 };
